error_handler (has_critical_error) mainly
